Modified Solution to Pull #179
- Fixed my weird white spacing
- Added a blank file for implementation
- Added @Ignore after First Test Case, I thought that would be good
practice
- Removed unused library util.Scanner from my first implementation
- Removed print statement
-Added Exercise to the config.json File
