+++ exercises/SumOfMultiples/src/test/SumOfMultiplesTest.java
@@ -1,92 +1,153 @@
 import static org.junit.Assert.*;
 
+import org.junit.Ignore;
 import org.junit.Test;
 
 public class SumOfMultiplesTest {
 
+	
 	@Test
-	public void test1() {
+	public void testSumOfMultiplesInputOne() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={3,5};
 		int output = mySum.Sum(1,set);
-		assertEquals(0,output);	
+		assertEquals(0,output);
+		
 	}
+	
+	
 	@Test
-    public void test2() {
+	@Ignore
+    public void testSumOfMultiplesInputFour() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={3,5};
 		int output = mySum.Sum(4,set);
 		assertEquals(3,output);
+		
 	}
+    
+    
 	@Test
-     public void test3() {
+	@Ignore
+     public void testSumOfMultiplesInputTen() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={3,5};
 		int output = mySum.Sum(10,set);
-		assertEquals(23,output);	
+		assertEquals(23,output);
+		
 	}
+     
+	
 	@Test
-     public void test4() {
+	@Ignore
+     public void testSumOfMultiplesInputOneHundred() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={3,5};
 		int output = mySum.Sum(100,set);
 		assertEquals(2318,output);
+		
 	}
+    
+	
 	@Test
-     public void test5() {
+	@Ignore
+     public void testSumOfMultiplesInputOneThousand() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={3,5};
 		int output = mySum.Sum(1000,set);
 		assertEquals(233168,output);
+		
 	}
+     
+	
 	@Test
-     public void test6() {
+	@Ignore
+     public void testSumOfMultiplesInputTwenty() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={7,13,17};
 		int output = mySum.Sum(20,set);
 		assertEquals(51,output);
+		
 	}
+     
+	
 	@Test
-     public void test7() {
+	@Ignore
+     public void testSumOfMultiplesInputThirty() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={4,6};
 		int output = mySum.Sum(15,set);
 		assertEquals(30,output);
+		
 	}
+     
+	
 	@Test
-     public void test8() {
+	@Ignore
+     public void testSumOfMultiplesInputOneHundredFifty() {
+		
 		SumOfMultiples mySum = new SumOfMultiples();
 		int[] set={5,6,8};
 		int output = mySum.Sum(150,set);
 		assertEquals(4419,output);
+		
 	}
 	
+	
 	@Test
-     public void test9() {
+	@Ignore
+     public void testSumOfMultiplesInputFiftyOne() {
+ 		
  		SumOfMultiples mySum = new SumOfMultiples();
  		int[] set={5,25};
  		int output = mySum.Sum(51,set);
  		assertEquals(275,output);
+ 		
  	}
+	
+	
 	@Test
-     public void test10() {
+	@Ignore
+     public void testSumOfMultiplesInputTenThousand() {
+  		
   		SumOfMultiples mySum = new SumOfMultiples();
   		int[] set={43,47};
   		int output = mySum.Sum(10000,set);
   		assertEquals(2203160,output);
+  		
   	}
+	
+	
 	@Test
-     public void test11() {
+	@Ignore
+     public void testSumOfMultiplesInputOneNumberArray() {
+   		
    		SumOfMultiples mySum = new SumOfMultiples();
    		int[] set={1};
    		int output = mySum.Sum(100,set);
-   		assertEquals(4950,output);	
+   		assertEquals(4950,output);
+   		
    	}
+	
+	
 	@Test
-     public void test12() {	
+	@Ignore
+     public void testSumOfMultiplesEmptyArray() {
+    		
     		SumOfMultiples mySum = new SumOfMultiples();
     		int[] set={};
     		int output = mySum.Sum(10000,set);
     		assertEquals(0,output);
+    		
     	}
+	
+	
+
 }
